requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.5
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
Django==4.1.3
django-admin-rangefilter==0.9.0
djangorestframework==3.14.0
drf-yasg==1.21.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kaitaistruct==0.10
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
pyasn1==0.4.8
pycparser==2.21
pydivert==2.1.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
selenium==4.6.0
selenium-wire==5.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
webdriver-manager==3.8.4
wsproto==1.2.0
zstandard==0.19.0
